Move types from Types.ts into packages
