+++ src/Types.ts
@@ -1,42 +1,11 @@
-// export type JSONType =
-//   | string
-//   | number
-//   | boolean
-//   | null
-//   | JSONObjectType
-//   | JSONArrayType;
-// export type JSONObjectType = { [key: string]: JSONType };
-// export type JSONArrayType = Array<string | number | boolean | Date | JSONType>;
-
+/**
+ * JSON
+ */
 export type JSONType = string | number | boolean | Date | null | JSONObjectType;
 
 export interface JSONObjectType {
   [x: string]: JSONType | JSONType[];
 }
-
-// interface JsonArray extends Array<string|number|boolean|Date|Json|JsonArray> { }
-
-export type SclangResultType = JSONType;
-
-export interface SynthDefResultType {
-  name: string;
-  bytes: Buffer;
-  synthDesc: {
-    sourceCode?: string;
-    // TODO: define the other values of synthDesc
-  };
-}
-export interface SynthDefResultMapType {
-  [defName: string]: SynthDefResultType;
-}
-
-interface SynthDefCompileRequestWithSource {
-  source: string;
-}
-interface SynthDefCompileRequestWithPath {
-  path: string;
-}
-export type SynthDefCompileRequest = SynthDefCompileRequestWithSource | SynthDefCompileRequestWithPath;
 
 /**
  * OSC
@@ -85,14 +54,4 @@ export interface BundleType {
 export interface CallAndResponse {
   call: MsgType;
   response: MsgType;
-}
-
-export interface NodeStateType {
-  parent?: number;
-  previous?: number;
-  next?: number;
-  isGroup: boolean;
-  head?: number;
-  tail?: number;
-  synthDef?: string;
 }
